Added a test of the Qt5 output
Connects (V4L2) input to a Qt5 output component, and uses the
MDI test widget to visualise the video. The raw frames can be
encoded in several ways (RGB, YUV) to test the shaders that are
used. Furthermore, the sourceID is artifically changed to simulate
sources being added and removed.