A/V Session chains can be started after 'init' - changed samples accordingly
This is useful in the case of the video session, so that a pointer
to the Qt output component can be obtained before starting the chain.
This way, the Qt signal indicating a new source is not missed.

The tests/qtouttest sample was changed to use an YUV file from disk
instead of V4L2 video